完成后台订单管理、用户信息、评论管理的列表显示

+++ src/main/java/org/lanqiao/dao/IUserDao.java
@@ -22,7 +22,7 @@ public interface IUserDao {
     //获取用户列表
     public List<User> selectUserList(Condition condition);
     //获取用户列表数量
-    public Long selectUserCount(Condition condition);
+    public int selectUserCount(Condition condition);
     //通过用户名密码查询用户
     public User selectUserByUserNameAndPassword(String username,String password);
     //通过用户名查询用户信息
